minimal test to reproduce #4
Put point anywhere within the SRC block and press C-c C-c. If the
result differs from this commit, or if you do not see two concentric
squares in different colors and containing the text "aaa" in a third
color, there is a regression from fbcb8b6 where this appears fixed.

IOW, If pressing C-c with the point somewhere in the emacs-lisp source code
doesn't work, or if the buffer content changes when you do, plz LMK