Fix #1: avoid double draw by checking the interned version of tags
The tiles parser was looking at the string rather than interned symbol
version of each tag when considering whether it was already defined
while loading symbols.   This each table which defined a tag for a
given symbol would cauase an additional version of that tag to be
added for that symbol, ultimate doubling all duplicate draw
instructions from each such tagged symbol.