click-to-reveal map cells
This is an hack so we can playtest via screen share hosted by the DM
who will prep the map quickly on a seperate unshared screen.

To view only the selected cells you must uncheck "view whole level" in
the Dungeon Map menu.

To save the visible tiles you must
 - M-x customize-options RET dm-map-current-level-cells RET
 - save for future sessions

Changing map levels is going to be a bit of a mess.