adding "dm-sketch", an interactive drawing feature
This is not currently hooked into emacs-init-scripts/user/init-dm.el

In theory, just:

  ;; fool with load-path
  (require 'dm-sketch)
  (dm-sketch)

This introduces dm-draw.el which begins a rewrite of the essencial
components of the SVG rendering functions from dm-map.el and also
dm-env, which provides a set of vars local across a paired set of
org-mode/image buffers.

A new hook `dm-character-mode-hook' is invoked when a character is
display including (when we support this) changing which character is
displayed.

`dm-character-mode' new major mode; however, few bindings will likely
work correctly at this point.

`dm-draw' is the main entry point for the new dm-draw.el rewrite of
dm-map.el "SVG rendering" functions.  This does not directly rely on
dm-env.el.

`dm-env-with' is the key function from the new dm-env.el.  This is
used by, e.g:

`dm-sketch' is the interactive drawing component.  Currently it's is
effectively hard-coded to give the "pencil" tool, we cannot change
line properties.  Only mouse bindings are defined to control drawing, Mouse-1 will add a
point, Mouse-2 will move-to a point.  A distinct visualization of the
line to be added (given a mouse-1 press) is shown.  There is no way to
add the newly created line/shape to the underlying drawing nor to save
or restore work.   Start the demo with (dm-sketch) it has been
`require'd.

The additional test draw-99 is aimed at quickly testing out new
features implemented in dm-draw and should not be mistaken for
meaningful test coverage.   All designs are POC and inputs welcome.